+++ SharkScope/Model/ApiRequestCache.swift
@@ -16,20 +16,18 @@ public struct ApiRequestCache
     public init()
     { }
     
-    func cachedResponse<ResponseType: Decodable>(for urlComponents: URLComponents) -> ResponseType?
+    func cachedResponse<RequestType: ApiRequest>(for request: RequestType) -> RequestType.ApiResponseType?
     {
         // Resolve file name.
-        guard let cacheFileURL = cacheFileURL(for: urlComponents)
+        guard let cacheFileURL = cacheFileURL(for: request)
         else { return nil }
         
         // Load / Decode JSON.
         do
         {
             let data = try Data(contentsOf: cacheFileURL)
-            let decoder = JSONDecoder()
-                decoder.keyDecodingStrategy = .convertFromBadgerFish
-                decoder.dateDecodingStrategy = .millisecondsSince1970
-            let response = try decoder.decode(ResponseType.self, from: data)
+            let string = String(decoding: data, as: UTF8.self)
+            let response = try RequestType.ApiResponseType.self(from: string)
             return response
         }
         catch
@@ -39,14 +37,14 @@ public struct ApiRequestCache
         }
     }
     
-    func cacheFolderURL(for urlComponents: URLComponents) -> URL?
+    func cacheFolderURL<RequestType: ApiRequest>(for request: RequestType) -> URL?
     {
         // Resolve Documents directory.
         guard let documentsDirectory = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask).first
-            else { return nil }
+        else { return nil }
         
-        // Parse subfolder (or fallback to no subfolder) without api path components.
-        let pathURL = URL(string: urlComponents.percentEncodedPath.replacingOccurrences(of: SharkScope.Service.basePath, with: "")) ?? URL(string: "")!
+        // Append path.
+        let pathURL = URL(string: request.path) ?? URL(string: "")!
         let pathFolder = pathURL.deletingLastPathComponent()
         let cacheFolderURL = documentsDirectory.appendingPathComponent(pathFolder.path, isDirectory: true)
         
@@ -57,11 +55,14 @@ public struct ApiRequestCache
         return cacheFolderURL
     }
     
-    func cacheFileURL(for urlComponents: URLComponents) -> URL?
+    func cacheFileURL<RequestType: ApiRequest>(for request: RequestType) -> URL?
     {
         // Try to resolve folder.
-        guard let cacheFolderURL = cacheFolderURL(for: urlComponents)
+        guard let cacheFolderURL = cacheFolderURL(for: request)
         else { return nil }
+        
+        // Encode parameters to query.
+        let urlComponents = request.urlComponents
         
         // Parse file name.
         let pathURL = URL(string: urlComponents.percentEncodedPath) ?? URL(string: "")!
@@ -71,8 +72,13 @@ public struct ApiRequestCache
         if urlComponents.percentEncodedQuery != ""
         { querySuffix = "?" + (urlComponents.percentEncodedQuery ?? "") }
         
+        // Determine extension.
+        var pathExtension: String = "json"
+        if (request.contentType == .CSV)
+        { pathExtension = "csv" }
+        
         // Assemble file name.
-        let cacheFileURL = cacheFolderURL.appendingPathComponent(pathURL.lastPathComponent + querySuffix).appendingPathExtension("json")
+        let cacheFileURL = cacheFolderURL.appendingPathComponent(pathURL.lastPathComponent + querySuffix).appendingPathExtension(pathExtension)
         
         return cacheFileURL
     }
@@ -82,18 +88,11 @@ public struct ApiRequestCache
     
     public func deleteTablesCache(for playerName: String)
     {
-        let requestPath = "activeTournaments"
-        let parameters = ["network1" : "PokerStars", "player1" : playerName]
+        // Create (fake) request.
+        let request = ActiveTournamentsRequest(network: "PokerStars", player: playerName)
         
-        // Create (fake) URL Components.
-        var urlComponents = URLComponents()
-            urlComponents.scheme = "https"
-            urlComponents.host = "sharkscope.com"
-            urlComponents.path = Service.basePath + requestPath
-            urlComponents.queryItems = parameters.map { eachElement in URLQueryItem(name: eachElement.key, value: eachElement.value) }
-    
         // Resolve file name.
-        guard let cacheFileURL = cacheFileURL(for: urlComponents)
+        guard let cacheFileURL = cacheFileURL(for: request)
         else { return }
         
         // Create folder (if needed).
@@ -103,19 +102,11 @@ public struct ApiRequestCache
     
     public func deleteStatisticsCache(for playerName: String)
     {
-        let network = "PokerStars"
-        let requestPath = "networks/\(network)/players/\(playerName)"
-        let parameters = PlayerSummaryRequest(network: network, player: playerName).parameters
+        // Create (fake) request.
+        let request = PlayerSummaryRequest(network: "PokerStars", player: playerName)
         
-        // Create (fake) URL Components.
-        var urlComponents = URLComponents()
-            urlComponents.scheme = "https"
-            urlComponents.host = "sharkscope.com"
-            urlComponents.path = SharkScope.Service.basePath + requestPath
-            urlComponents.queryItems = parameters.map { eachElement in URLQueryItem(name: eachElement.key, value: eachElement.value) }
-    
         // Resolve file name.
-        guard let cacheFileURL = cacheFileURL(for: urlComponents)
+        guard let cacheFileURL = cacheFileURL(for: request)
         else { return }
         
         // Create folder (if needed).